requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
numpy==1.19.1
owotranslator==0.0.6
Pillow==7.2.0
praw==7.1.0
prawcore==1.5.0
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
typing-extensions==3.7.4.2
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
wincertstore==0.2
yarl==1.5.1
